requirements.txt
flask
torch
torchvision
ultralytics
Pillow
opencv-python